+++ changelog.md
@@ -9,6 +9,9 @@ Dates are in YYYY/MM/DD format, time is in America/Chicago (CDT/CST), UTC-0500/-
 2023/01/16-2023/07/17
 
 ### v0.0.24 (2023/07/17 01:XX)
+Fixes:
+* `Server` decorators now have correct method names in docstrings.
+
 Misc.:
 * `Server.stream` is now decorated with `@Server._make_method`