+++ changelog.md
@@ -8,7 +8,7 @@ Dates are in YYYY/MM/DD format, time is in America/Chicago (CDT/CST), UTC-0500/-
 ## Developmental (0.0.X)
 2023/01/16-2023/07/17
 
-### v0.0.24 (2023/07/17 01:XX)
+### v0.0.24 (2023/07/17 17:41)
 Features:
 * Native GraphQL support! Use `@server.gql(schema:str, endpoint:str)`.